+++ src/github/github.ts
@@ -148,7 +148,7 @@ export class Github {
    */
   removeUnwantedFiles(outDir: string) {
     fileFilterList.forEach((filename) => {
-        rimraf.sync(path.join(outDir, filename));
+      rimraf.sync(path.join(outDir, filename));
     });
   }
 
@@ -172,10 +172,8 @@ export class Github {
     const maxSatisfyingRelease =
         releases.find((r) => r.tag_name === maxSatisfyingReleaseVersion);
     if (!maxSatisfyingRelease) {
-      throw new Error(
-          `${this._owner}/${this._repo} has no releases matching ${
-                                                                   semverRange
-                                                                 }.`);
+      throw new Error(`${this._owner}/${this._repo} has no releases matching ${
+          semverRange}.`);
     }
     return maxSatisfyingRelease;
   }